Initialisation
I just created the html and css file. At first, I created the different sections of the website and defined their backgroud color, margin, padding and font family and font size. That was the goal for today, tomorrow we will work on the first grid areas and then the navigation bar